Merging menu changes from master and my recent stuff
Recent stuff, added the wall and the health bar. 
When enemies get to the wall they stop and attack it. There is a counter between attacks so atm they attack every 1.5 seconds and do 1 damage. The health bar looks bad, will need to get fiddled with later. 
When the health gets to 0 a 'game over' text appears and then the game closes within 3 seconds.